rename: clarify intent
There may be more than just the bepost adapter.  Let's put
adapters in folders starting with `adapter_`.
