+++ src/finhan/adapter_bepost/test_csv_transactions.py
@@ -2,7 +2,7 @@ from datetime import datetime
 
 import pytest
 
-from .bepost import from_lines, Transaction
+from .csv_transactions import from_lines, Transaction
 
 SAMPLE = """Rekeningnummer :;BE0123456789;b.compact rekening
 Nummer van de verrichting;Transactie datum;Beschrijving;Bedrag van de verrichting;Munt;Valuta datum;Rekening tegenpartij :;Naam tegenpartij;Mededeling;Referentie van de verrichting
